Removed stedcj from the public API
