+++ library/include/rocsolver/rocsolver-functions.h
@@ -3976,92 +3976,6 @@ ROCSOLVER_EXPORT rocblas_status rocsolver_zstedc(rocblas_handle handle,
 //! @}
 
 /*! @{
-    \brief STEDCJ computes the eigenvalues and (optionally) eigenvectors of
-    a symmetric tridiagonal matrix.
-
-    \details
-    This function uses the divide and conquer method to compute the eigenvectors.
-    The eigenvalues are returned in increasing order.
-
-    The matrix is not represented explicitly, but rather as the array of
-    diagonal elements D and the array of symmetric off-diagonal elements E.
-    When D and E correspond to the tridiagonal form of a full symmetric/Hermitian matrix, as returned by, e.g.,
-    \ref rocsolver_ssytrd "SYTRD" or \ref rocsolver_chetrd "HETRD", the eigenvectors of the original matrix can also
-    be computed, depending on the value of evect.
-
-    @param[in]
-    handle      rocblas_handle.
-    @param[in]
-    evect       #rocblas_evect.
-                Specifies how the eigenvectors are computed.
-    @param[in]
-    n           rocblas_int. n >= 0.
-                The number of rows and columns of the tridiagonal matrix.
-    @param[inout]
-    D           pointer to real type. Array on the GPU of dimension n.
-                On entry, the diagonal elements of the tridiagonal matrix.
-                On exit, if info = 0, the eigenvalues in increasing order.
-    @param[inout]
-    E           pointer to real type. Array on the GPU of dimension n-1.
-                On entry, the off-diagonal elements of the tridiagonal matrix.
-                On exit, if info = 0, the values of this array are destroyed.
-    @param[inout]
-    C           pointer to type. Array on the GPU of dimension ldc*n.
-                On entry, if evect is original, the orthogonal/unitary matrix
-                used for the reduction to tridiagonal form as returned by, e.g.,
-                \ref rocsolver_sorgtr "ORGTR" or \ref rocsolver_cungtr "UNGTR".
-                On exit, if info = 0, it is overwritten with the eigenvectors of the original
-                symmetric/Hermitian matrix (if evect is original), or the
-                eigenvectors of the tridiagonal matrix (if evect is tridiagonal).
-                (Not referenced if evect is none).
-    @param[in]
-    ldc         rocblas_int. ldc >= n if evect is original or tridiagonal.
-                Specifies the leading dimension of C. (Not referenced if evect is none).
-    @param[out]
-    info        pointer to a rocblas_int on the GPU.
-                If info = 0, successful exit.
-                If info = i > 0, STEDCJ failed to compute an eigenvalue on the sub-matrix formed by
-                the rows and columns info/(n+1) through mod(info,n+1).
-    ********************************************************************/
-
-ROCSOLVER_EXPORT rocblas_status rocsolver_sstedcj(rocblas_handle handle,
-                                                  const rocblas_evect evect,
-                                                  const rocblas_int n,
-                                                  float* D,
-                                                  float* E,
-                                                  float* C,
-                                                  const rocblas_int ldc,
-                                                  rocblas_int* info);
-
-ROCSOLVER_EXPORT rocblas_status rocsolver_dstedcj(rocblas_handle handle,
-                                                  const rocblas_evect evect,
-                                                  const rocblas_int n,
-                                                  double* D,
-                                                  double* E,
-                                                  double* C,
-                                                  const rocblas_int ldc,
-                                                  rocblas_int* info);
-
-ROCSOLVER_EXPORT rocblas_status rocsolver_cstedcj(rocblas_handle handle,
-                                                  const rocblas_evect evect,
-                                                  const rocblas_int n,
-                                                  float* D,
-                                                  float* E,
-                                                  rocblas_float_complex* C,
-                                                  const rocblas_int ldc,
-                                                  rocblas_int* info);
-
-ROCSOLVER_EXPORT rocblas_status rocsolver_zstedcj(rocblas_handle handle,
-                                                  const rocblas_evect evect,
-                                                  const rocblas_int n,
-                                                  double* D,
-                                                  double* E,
-                                                  rocblas_double_complex* C,
-                                                  const rocblas_int ldc,
-                                                  rocblas_int* info);
-//! @}
-
-/*! @{
     \brief STEBZ computes a set of eigenvalues of a symmetric tridiagonal matrix T.
 
     \details
